fix: fix api

+++ backend/controllers/cards.js
@@ -10,14 +10,21 @@ module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
 
   Card.create({ name, link, owner: req.user._id })
-    .then((card) => res.send(card))
+    .then((card) => res.status(201).send(card))
     .catch(next);
 };
 
 module.exports.deleteCard = (req, res, next) => {
-  Card.findByIdAndRemove(req.params.cardId)
+  Card.findById(req.params.cardId)
     .orFail(new Error('notValidId'))
-    .then((card) => res.send(card))
+    .then((data) => {
+      // eslint-disable-next-line eqeqeq
+      if (data.owner != req.user._id) {
+        throw new Error('forbidden');
+      }
+      Card.findByIdAndRemove(req.params.cardId)
+        .then((card) => res.send(card));
+    })
     .catch(next);
 };
